wip: create process-test-directory folder and utils.js
