+++ .eslintignore
@@ -1,4 +1,4 @@
 tests/**/reference/**/*
 build/tests/**/*
-# keys are manually parsed in process-test-directory.js and depend on format
+# keys are manually parsed in process-test-directory and depend on format
 tests/resources/keys.mjs
