+++ lib/data/process-test-directory/utils.js
@@ -1,0 +1,122 @@
+const {
+  types: { isArrayBufferView, isArrayBuffer },
+} = require('util');
+const { reindent } = require('../../util/lines');
+
+const logger = (isVerbose, isValidate) => {
+  // setup from arguments passed to npm script
+  const VERBOSE_CHECK = isVerbose;
+  const VALIDATE_CHECK = isValidate;
+
+  let suppressedMessages = 0;
+
+  /**
+   * @param {string} message - message to be logged
+   * @param {object} [options]
+   * @param {boolean} [options.severe=false] - indicates whether the message should be viewed as an error or not
+   * @param {boolean} [options.force=false] - indicates whether this message should be forced to be outputted regardless of verbosity level
+   */
+  const log = (message, { severe = false, force = false } = {}) => {
+    if (VERBOSE_CHECK || force) {
+      if (severe) console.error(message);
+      else console.log(message);
+    } else {
+      // Output no logs
+      suppressedMessages += 1; // counter to indicate how many messages were hidden
+    }
+  };
+
+  /**
+   * @param {string} message - error message
+   */
+  log.warning = message => log(message, { severe: true, force: true });
+
+  /**
+   * Log error then exit the process.
+   * @param {string} message - error message
+   */
+  log.error = message => {
+    log.warning(message);
+    process.exit(1);
+  };
+
+  log.suppressedMessages = () => suppressedMessages;
+
+  return { log, VALIDATE_CHECK };
+};
+
+// Filter out reference html files with inline scripts. Files that are not
+// regenerated will be removed from the filesystem.
+const testPlanUpdate = async testPlanRecord =>
+  await testPlanRecord.walk(record => {
+    if (record.entries) {
+      return {
+        ...record,
+        entries: record.entries.filter(record => !isScriptedReferenceRecord(record)),
+      };
+    }
+    return record;
+  });
+
+function isScriptedReferenceRecord(subrecord) {
+  return (
+    subrecord.name &&
+    subrecord.name.endsWith('.html') &&
+    subrecord.name.indexOf('.') !== subrecord.name.lastIndexOf('.')
+  );
+}
+
+function createExampleScriptedFile(name, examplePathTemplate, exampleTemplate, source) {
+  return {
+    name,
+    path: examplePathTemplate(name),
+    content: exampleTemplate.render(exampleTemplateParams(name, source)).toString(),
+  };
+}
+
+function toBuffer(content) {
+  if (Buffer.isBuffer(content) || isArrayBufferView(content) || isArrayBuffer(content)) {
+    return content;
+  } else if (typeof content === 'string') {
+    return Buffer.from(content);
+  }
+  return Buffer.from(content.toString());
+}
+
+function exampleTemplateParams(name, source) {
+  return {
+    script: reindent`
+<!-- Generated by process-test-directory.js -->
+<script>
+  (function() {
+    function setupScript(testPageDocument) {
+      // ${name}
+      ${source}
+    };
+    document.addEventListener('click', function(event) {
+      if (event.target.classList.contains('button-run-test-setup')) {
+        event.target.disabled = true;
+        setupScript(document);
+      }
+    });
+  })();
+</script>
+<!-- End of generated output -->`,
+    button: reindent`
+<!-- Generated by process-test-directory.js -->
+<div style="position: relative; left: 0; right: 0; height: 2rem;">
+  <button class="button-run-test-setup" autofocus style="height: 100%; width: 100%;"${
+    source ? '' : ' disabled'
+  }>Run Test Setup</button>
+</div>
+<!-- End of generated output -->`,
+  };
+}
+
+module.exports = {
+  logger,
+  testPlanUpdate,
+  isScriptedReferenceRecord,
+  createExampleScriptedFile,
+  toBuffer,
+};
