+++ lib/data/process-test-directory/v1.js
@@ -1,76 +1,42 @@
-/// <reference path="../../types/aria-at-csv.js" />
-/// <reference path="../../types/aria-at-parsed.js" />
-/// <reference path="../../types/aria-at-validated.js" />
-/// <reference path="../../types/aria-at-file.js" />
-/// <reference path="../util/file-record-types.js" />
+/// <reference path="../../../types/aria-at-csv.js" />
+/// <reference path="../../../types/aria-at-parsed.js" />
+/// <reference path="../../../types/aria-at-validated.js" />
+/// <reference path="../../../types/aria-at-file.js" />
+/// <reference path="../../util/file-record-types.js" />
 
 'use strict';
 
 const path = require('path');
 const { Readable } = require('stream');
-const {
-  types: { isArrayBufferView, isArrayBuffer },
-} = require('util');
 
 const csv = require('csv-parser');
 const beautify = require('json-beautify');
 
-const { validate } = require('../util/error');
-const { reindent } = require('../util/lines');
-const { Queryable } = require('../util/queryable');
-const { FileRecordChain } = require('../util/file-record-chain');
+const { validate } = require('../../util/error');
+const { reindent } = require('../../util/lines');
+const { Queryable } = require('../../util/queryable');
+const { FileRecordChain } = require('../../util/file-record-chain');
 
-const { parseSupport } = require('./parse-support');
-const { parseTestCSVRow } = require('./parse-test-csv-row');
-const { parseCommandCSVRow } = require('./parse-command-csv-row');
-const { createCommandTuplesATModeTaskLookup } = require('./command-tuples-at-mode-task-lookup');
+const { parseSupport } = require('../parse-support');
+const { parseTestCSVRow } = require('../parse-test-csv-row');
+const { parseCommandCSVRow } = require('../parse-command-csv-row');
+const { createCommandTuplesATModeTaskLookup } = require('../command-tuples-at-mode-task-lookup');
 
-const { renderHTML: renderCollectedTestHtml } = require('./templates/collected-test.html');
-const { createExampleScriptsTemplate } = require('./example-scripts-template');
+const { renderHTML: renderCollectedTestHtml } = require('../templates/collected-test.html');
+const { createExampleScriptsTemplate } = require('../example-scripts-template');
+const {
+  logger,
+  isScriptedReferenceRecord,
+  createExampleScriptedFile,
+  toBuffer,
+} = require('./utils');
 
 /**
  * @param {string} directory - path to directory of data to be used to generate test
  * @param {object} args={}
  */
 const processTestDirectory = async ({ directory, args = {} }) => {
-  let VERBOSE_CHECK = false;
-  let VALIDATE_CHECK = false;
-
-  let suppressedMessages = 0;
-
-  /**
-   * @param {string} message - message to be logged
-   * @param {object} [options]
-   * @param {boolean} [options.severe=false] - indicates whether the message should be viewed as an error or not
-   * @param {boolean} [options.force=false] - indicates whether this message should be forced to be outputted regardless of verbosity level
-   */
-  const log = (message, { severe = false, force = false } = {}) => {
-    if (VERBOSE_CHECK || force) {
-      if (severe) console.error(message);
-      else console.log(message);
-    } else {
-      // Output no logs
-      suppressedMessages += 1; // counter to indicate how many messages were hidden
-    }
-  };
-
-  /**
-   * @param {string} message - error message
-   */
-  log.warning = message => log(message, { severe: true, force: true });
-
-  /**
-   * Log error then exit the process.
-   * @param {string} message - error message
-   */
-  log.error = message => {
-    log.warning(message);
-    process.exit(1);
-  };
-
-  // setup from arguments passed to npm script
-  VERBOSE_CHECK = !!args.verbose;
-  VALIDATE_CHECK = !!args.validate;
+  const { log, VALIDATE_CHECK } = logger(!!args.verbose, !!args.validate);
 
   const validModes = ['reading', 'interaction', 'item'];
 
@@ -79,7 +45,7 @@ const processTestDirectory = async ({ directory, args = {} }) => {
 
   // cwd; @param rootDirectory is dependent on this file not moving from the `lib/data` folder
   const libDataDirectory = path.dirname(__filename);
-  const rootDirectory = path.join(libDataDirectory, '../..');
+  const rootDirectory = path.join(libDataDirectory, '../../..');
 
   const testsDirectory = path.join(rootDirectory, 'tests');
   const testPlanDirectory = path.join(rootDirectory, 'tests', testPlanName);
@@ -881,7 +847,7 @@ ${rows}
     log('No validation errors detected\n');
   }
 
-  return { isSuccessfulRun: errorCount === 0, suppressedMessages };
+  return { isSuccessfulRun: errorCount === 0, suppressedMessages: log.suppressedMessages() };
 };
 
 exports.processTestDirectory = processTestDirectory;
@@ -1338,60 +1304,5 @@ function createCollectedTestHtmlFile(test, testPlanBuildDirectory) {
       )}-${test.target.mode}-${test.target.at.key}.collected.html`
     ),
     content: encodeText(renderCollectedTestHtml(test, testJsonFileName)),
-  };
-}
-
-function isScriptedReferenceRecord(subrecord) {
-  return (
-    subrecord.name &&
-    subrecord.name.endsWith('.html') &&
-    subrecord.name.indexOf('.') !== subrecord.name.lastIndexOf('.')
-  );
-}
-
-function createExampleScriptedFile(name, examplePathTemplate, exampleTemplate, source) {
-  return {
-    name,
-    path: examplePathTemplate(name),
-    content: exampleTemplate.render(exampleTemplateParams(name, source)).toString(),
-  };
-}
-
-function toBuffer(content) {
-  if (Buffer.isBuffer(content) || isArrayBufferView(content) || isArrayBuffer(content)) {
-    return content;
-  } else if (typeof content === 'string') {
-    return Buffer.from(content);
-  }
-  return Buffer.from(content.toString());
-}
-
-function exampleTemplateParams(name, source) {
-  return {
-    script: reindent`
-<!-- Generated by process-test-directory.js -->
-<script>
-  (function() {
-    function setupScript(testPageDocument) {
-      // ${name}
-      ${source}
-    };
-    document.addEventListener('click', function(event) {
-      if (event.target.classList.contains('button-run-test-setup')) {
-        event.target.disabled = true;
-        setupScript(document);
-      }
-    });
-  })();
-</script>
-<!-- End of generated output -->`,
-    button: reindent`
-<!-- Generated by process-test-directory.js -->
-<div style="position: relative; left: 0; right: 0; height: 2rem;">
-  <button class="button-run-test-setup" autofocus style="height: 100%; width: 100%;"${
-    source ? '' : ' disabled'
-  }>Run Test Setup</button>
-</div>
-<!-- End of generated output -->`,
   };
 }
